+++ test_problems/stoichSolidKinetics/stoichSolidKinetics.cpp
@@ -70,9 +70,9 @@ void testProblem()
     importKinetics(*xg, phaseList, &iKin);
 
     vector_fp mll(gasTP->nSpecies(), 0.0);
-    int igco2 = gasTP->speciesIndex("CO2");
-    int igo2  = gasTP->speciesIndex("O2");
-    int ign2  = gasTP->speciesIndex("N2");
+    size_t igco2 = gasTP->speciesIndex("CO2");
+    size_t igo2  = gasTP->speciesIndex("O2");
+    size_t ign2  = gasTP->speciesIndex("N2");
 
     mll[igco2] = 0.2;
     mll[igo2] = 0.1;
@@ -131,13 +131,13 @@ void testProblem()
         printRates(iKin);
 
         cout << "*** Setting CaO(S) phase to nonexistent:" << endl;
-        int ip_cao = iKin.phaseIndex("CaO(S)");
+        size_t ip_cao = iKin.phaseIndex("CaO(S)");
         iKin.setPhaseExistence(ip_cao, false);
         iKin.setPhaseStability(ip_cao, true);
         printRates(iKin);
 
         cout << "*** Setting CaCO3(S) phase to nonexistent:" << endl;
-        int ip_caco3 = iKin.phaseIndex("CaCO3(S)");
+        size_t ip_caco3 = iKin.phaseIndex("CaCO3(S)");
         iKin.setPhaseExistence(ip_caco3, false);
         iKin.setPhaseStability(ip_caco3, true);
         printRates(iKin);
@@ -147,7 +147,7 @@ void testProblem()
         printRates(iKin);
 
         cout << "*** Setting Gas phase to nonexistent, CaCO3 nonexistent:" << endl;
-        int ip_gas = iKin.phaseIndex("air");
+        size_t ip_gas = iKin.phaseIndex("air");
         iKin.setPhaseExistence(ip_gas, false);
         iKin.setPhaseStability(ip_gas, true);
         printRates(iKin);
